+++ src/io/CompressedOutputStream.cpp
@@ -783,8 +783,8 @@ T EncodingTask<T>::run()
                 // Create message (use snprintf because stringstream is too slow)
                 char buf2[100];
                 snprintf(buf2, sizeof(buf2),
-                         "{ \"type\":\"%s\", \"id\":%d, \"offset\":%lu, \"skipFlags\":%s }",
-                         "BLOCK_INFO", blockId, blockOffset, buf1);
+                         "{ \"type\":\"%s\", \"id\":%d, \"offset\":%llu, \"skipFlags\":%s }",
+                         "BLOCK_INFO", blockId, (unsigned long long)blockOffset, buf1);
                 Event evt2(Event::BLOCK_INFO, blockId, string(buf2));
                 CompressedOutputStream::notifyListeners(_listeners, evt2);
             }
